Add custom metrics query argument
This commit adds support for a `custom_metrics_query` argument. The
argument must be a SQL query string that has the columns 'metric_name',
'metric_value', and 'metric_type'. Additional columns are not required,
but if they exist, they will be added to the metric set as attributes
where the column name is the attribute key and the value for that row is the
attribute value.