custom query update (#76)
* custom query update

This update expands on what was in the sample file before. This new set of custom queries are already in use by some of our customers today. This PR just makes it officially part of our repo. There is also a corresponding dashboard in quickstarts that will be checked in once this is merged.